typescript/user-interface/packages/ui-state/src/ts/workers/waveform-worker/operations/ui-filter-processor.ts
import type { ChannelSegmentTypes, FilterTypes, WaveformTypes } from '@gms/common-model';
import type { FilterDefinition } from '@gms/common-model/lib/filter/types';
import { design, filter } from '@gms/ui-wasm';

import type { SampleRate } from '../../../types';
import { changeUniqueIdFilter, storePositionBuffer } from '../util/position-buffer-util';
import { WaveformStore } from '../worker-store';

// !NOTE: the data is in a format like [time,value,time,value,...] set the index offset appropriately
const indexOffset = 1; // start with the first y value of the data
const indexInc = 2; // have the algorithm skip each x value of the data (only filter y values)

/**
 * Defines the parameters for the {@link designFilter} operation.
 */
interface DesignFilterParams {
  filterDefinition: FilterTypes.FilterDefinition;
  taper: number;
  removeGroupDelay: boolean;
}

/**
 * Designs a Filter Definition by populating the coefficients
 *
 * @param filterDefinition the filter definition to design
 * @param taper number of samples for cosine taper
 * @param removeGroupDelay optional boolean to determine if group delay should be applied
 * @returns the designed filter definition
 */
export const designFilter = async ({
  filterDefinition
}: DesignFilterParams): Promise<FilterTypes.FilterDefinition> =>
  Promise.resolve(design(filterDefinition));

const filterDataSegments = async (
  timeseriesList: WaveformTypes.Waveform[],
  filterDefinitions: Record<SampleRate, FilterDefinition>,
  taper: number,
  removeGroupDelay: boolean
): Promise<WaveformTypes.Waveform[]> => {
  return Promise.all(
    timeseriesList.map(async (timeseries: WaveformTypes.Waveform) => {
      if (timeseries._uiClaimCheckId) {
        // Get wave from store
        const wave = await WaveformStore.retrieve(timeseries._uiClaimCheckId);
        // Apply filter to wave
        const filterPromise = filter(
          filterDefinitions[timeseries.sampleRateHz],
          wave,
          taper,
          removeGroupDelay,
          indexOffset,
          indexInc
        );

        // store the new filtered data in the waveform cache
        const filteredId = changeUniqueIdFilter(
          timeseries._uiClaimCheckId,
          filterDefinitions[timeseries.sampleRateHz].name
        );
        // TODO: Try parallelizing this request by storing the promise for the filter operation rather than the filtered value
        await storePositionBuffer(filteredId, filterPromise);

        return {
          ...timeseries,
          // update the claim check id for the new data segment
          _uiClaimCheckId: filteredId
        };
      }
      throw new Error('Filter processor filter operation was passed invalid channelSegment data');
    })
  );
};

/**
 * Defines the parameters for the {@link filterChannelSegment} operation.
 */
interface FilterChannelSegmentParams {
  channelSegment: ChannelSegmentTypes.ChannelSegment<WaveformTypes.Waveform>;
  filterDefinitions: Record<SampleRate, FilterDefinition>;
  taper: number;
  removeGroupDelay: boolean;
}

export const filterChannelSegment = async ({
  channelSegment,
  filterDefinitions,
  taper,
  removeGroupDelay
}: FilterChannelSegmentParams): Promise<
  ChannelSegmentTypes.ChannelSegment<WaveformTypes.Waveform>
> => {
  const filterDefinitionName = Object.values(filterDefinitions)?.[0]?.name;

  const timeseries = await filterDataSegments(
    channelSegment.timeseries,
    filterDefinitions,
    taper,
    removeGroupDelay
  );
  return {
    ...channelSegment,
    _uiFilterId: filterDefinitionName,
    timeseries
  };
};

/**
 * Defines the parameters for the {@link filterChannelSegments} operation.
 */
interface FilterChannelSegmentsParams {
  channelSegments: ChannelSegmentTypes.ChannelSegment<WaveformTypes.Waveform>[];
  filterDefinitions: Record<SampleRate, FilterDefinition>;
  taper: number;
  removeGroupDelay: boolean;
}

/**
 * Applies a Filter Definition to the provided channel segments (filters the data).
 *
 * @param channelSegments the waveform channel segments
 * @param filterDefinition a Linear Filter Definition
 * @param taper number of samples for cosine taper
 * @param removeGroupDelay optional boolean to determine if group delay should be applied
 * @returns the filtered waveform channel segments
 */
export const filterChannelSegments = async ({
  channelSegments,
  filterDefinitions,
  taper,
  removeGroupDelay
}: FilterChannelSegmentsParams): Promise<
  ChannelSegmentTypes.ChannelSegment<WaveformTypes.Waveform>[]
> => {
  return Promise.all(
    channelSegments.map(async channelSegment => {
      return filterChannelSegment({
        channelSegment,
        filterDefinitions,
        taper,
        removeGroupDelay
      });
    })
  );
};
